+++ DosingApp/DosingApp/Models/Assignment.cs
@@ -65,7 +65,7 @@ namespace DosingApp.Models
         public int? FieldId { get; set; }
         public virtual Field Field { get; set; }
 
-        public float? VolumeRate { get; set; }
+        public double? VolumeRate { get; set; }
         public string Unit { get; set; }
     }
 }
